+++ examples/chat-clerk/src/schema.ts
@@ -1,0 +1,8 @@
+import { CoMap, CoList, co } from "jazz-tools";
+
+
+export class Message extends CoMap {
+    text = co.string;
+}
+
+export class Chat extends CoList.Of(co.ref(Message)) {}