+++ examples/chat/src/app.tsx
@@ -1,20 +1,11 @@
-import { CoMap, CoList, co, Group, ID } from "jazz-tools";
-import { createJazzReactApp, DemoAuthBasicUI, useDemoAuth } from "jazz-react";
-import { createRoot } from "react-dom/client";
+import { Group, ID } from "jazz-tools";
 import { useIframeHashRouter } from "hash-slash";
+import { useAccount } from "./main.tsx";
+import { Chat } from "./schema.ts";
 import { ChatScreen } from "./chatScreen.tsx";
-import { StrictMode } from "react";
+import { AppContainer, TopBar } from "./dumbComponents.tsx";
 
-export class Message extends CoMap {
-    text = co.string;
-}
-
-export class Chat extends CoList.Of(co.ref(Message)) {}
-
-const Jazz = createJazzReactApp();
-export const { useAccount, useCoState } = Jazz;
-
-function App() {
+export function App() {
     const { me, logOut } = useAccount();
 
     const createChat = () => {
@@ -26,39 +17,14 @@ function App() {
     };
 
     return (
-        <div className="flex flex-col items-center justify-between w-screen h-screen p-2 dark:bg-black dark:text-white">
-            <div className="rounded mb-5 px-2 py-1 text-sm self-end">
-                {me?.profile?.name} ·{" "}
-                <button onClick={logOut}>Log Out</button>
-            </div>
+        <AppContainer>
+            <TopBar>
+                {me?.profile?.name} · <button onClick={logOut}>Log Out</button>
+            </TopBar>
             {useIframeHashRouter().route({
                 "/": () => createChat() as never,
                 "/chat/:id": (id) => <ChatScreen chatID={id as ID<Chat>} />,
             })}
-        </div>
+        </AppContainer>
     );
 }
-
-function AuthAndJazz({ children }: { children: React.ReactNode }) {
-    const [auth, state] = useDemoAuth();
-
-    return (
-        <>
-            <Jazz.Provider
-                auth={auth}
-                peer="wss://mesh.jazz.tools/?key=chat-example-jazz@gcmp.io"
-            >
-                {children}
-            </Jazz.Provider>
-            <DemoAuthBasicUI appName="Jazz Chat" state={state} />
-        </>
-    );
-}
-
-createRoot(document.getElementById("root")!).render(
-    <StrictMode>
-        <AuthAndJazz>
-            <App />
-        </AuthAndJazz>
-    </StrictMode>
-);